+++ packages/mechanic/src/mechanic-validation.js
@@ -12,66 +12,75 @@ const otherParams = { preset: true, scaleToFit: true, randomSeed: true };
  */
 const validateParams = params => {
   // Check all params
-  for (let param in params) {
+  for (let paramName in params) {
+    const param = params[paramName];
+    const { type: paramType, default: paramDefault } = param;
+
     // Check for all required properties
     for (let requiredKey of requiredKeys) {
-      if (!hasKey(params[param], requiredKey)) {
-        return `Parameter ${param} must have '${requiredKey}' property.`;
+      if (!hasKey(param, requiredKey)) {
+        return `Parameter ${paramName} must have '${requiredKey}' property.`;
       }
     }
 
     // Check type is supported by Mechanic
-    if (!hasKey(supportedTypes, params[param].type)) {
-      return `Parameter of type ${params[param].type} not supported, expected: ${Object.keys(
+    if (!hasKey(supportedTypes, paramType)) {
+      return `Parameter of type ${paramType} not supported, expected either: ${Object.keys(
         supportedTypes
       )}.`;
     }
 
     // Check default value is of correct type to corresponding param type
-    if (typeof params[param].default !== supportedTypes[params[param].type]) {
-      return `Default property value invalid for parameter ${param} of type ${
-        params[param].type
-      }. Expected to be ${supportedTypes[params[param].type]}.`;
+    if (typeof paramDefault !== supportedTypes[paramType]) {
+      return `Default property value invalid for parameter ${paramName} of type ${paramType}. Expected to be ${supportedTypes[paramType]}.`;
     }
 
     // Check that 'validation' property is function
-    if (hasKey(params[param], "validation") && typeof params[param].validation !== "function") {
-      return `Expected function in validation property in ${param}.`;
+    if (hasKey(param, "validation") && typeof param.validation !== "function") {
+      return `Expected function in validation property in ${paramName}. Got ${typeof param.validation}.`;
     }
 
-    // TODO: Add "Check that 'when' property is function"
+    // Check that 'editable' property is function
+    if (
+      hasKey(param, "editable") &&
+      typeof param.editable !== "function" &&
+      typeof param.editable !== "boolean"
+    ) {
+      return `Expected function or boolean in editable property in ${paramName}. Got ${typeof param.editable}.`;
+    }
 
-    // Check that 'options' property
-    if (hasKey(params[param], "options")) {
+    // Check 'options' property
+    if (hasKey(param, "options")) {
+      const { options } = param;
       // Should be function or object
-      if (!Array.isArray(params[param].options) && typeof params[param].options !== "object") {
-        return `Expected array or object in options property in ${param}. Received ${typeof params[
-          param
-        ].options}`;
+      if (!Array.isArray(options) && typeof options !== "object") {
+        return `Expected array or object in options property in ${paramName}. Got ${typeof options}.`;
       }
       // If it's not an array, it's an object
-      else if (!Array.isArray(params[param].options)) {
+      else if (!Array.isArray(options)) {
         // It should be consistent with 'default' property
-        if (!Object.keys(params[param].options).includes(params[param].default)) {
-          return `Default value ${params[param].default} for ${param} is not present in given options. `;
+        if (!Object.keys(options).includes(paramDefault)) {
+          return `Default value ${paramDefault} for ${paramName} is not present in given options: ${Object.keys(
+            options
+          )}. `;
         }
         // All values should be consistent with the param type
-        for (let option in params[param].options) {
-          if (typeof params[param].options[option] !== supportedTypes[params[param].type]) {
-            return `Incorrect type of value in options object. Expected ${params[param].type}.`;
+        for (let option in options) {
+          if (typeof options[option] !== supportedTypes[paramType]) {
+            return `Incorrect type of value in options object. Expected ${paramType}.`;
           }
         }
       }
       // If it's an array
-      else if (Array.isArray(params[param].options)) {
+      else if (Array.isArray(options)) {
         // It should be consistent with 'default' property
-        if (!params[param].options.includes(params[param].default)) {
-          return `Default value ${params[param].default} for ${param} is not present in given options. `;
+        if (!options.includes(paramDefault)) {
+          return `Default value ${paramDefault} for ${paramName} is not present in given options: ${options} `;
         }
         // All values should be consistent with the param type
-        for (let option of params[param].options) {
-          if (typeof option !== supportedTypes[params[param].type]) {
-            return `Incorrect type of value in options array. Expected ${params[param].type}.`;
+        for (let option of options) {
+          if (typeof option !== supportedTypes[paramType]) {
+            return `Incorrect type of value ${option} (${typeof option}) in options array. Expected ${paramType}.`;
           }
         }
       }
@@ -87,33 +96,32 @@ const validateParams = params => {
  * @param {object} values - Values for some or all of the parameters
  */
 const validateValues = (params, values = {}) => {
-  for (let param in params) {
+  for (let paramName in params) {
+    const param = params[paramName];
     // Validate that values for options are specified in the template
-    if (hasKey(values, param) && hasKey(params[param], "options")) {
-      if (Array.isArray(params[param].options) && !params[param].options.includes(values[param])) {
-        return `Supplied ${param} parameter is not available in the template options: ${params[param].options}`;
-      } else if (
-        !Array.isArray(params[param].options) &&
-        !hasKey(params[param].options, values[param])
-      ) {
-        return `Supplied ${param} parameter (${
-          values[param]
-        }) is not available in the template options: ${Object.keys(params[param].options)}`;
+    if (hasKey(values, paramName) && hasKey(param, "options")) {
+      const { options } = param;
+      if (Array.isArray(options) && !options.includes(values[paramName])) {
+        return `Supplied ${paramName} parameter is not available in the template options: ${options}`;
+      } else if (!Array.isArray(options) && !hasKey(options, values[paramName])) {
+        return `Supplied ${paramName} parameter (${
+          values[paramName]
+        }) is not available in the template options: ${Object.keys(options)}`;
       }
     }
     // Run validation functions for values.
-    if (hasKey(values, param) && hasKey(params[param], "validation")) {
-      const error = params[param].validation(values[param]);
-      if (error !== null) {
-        return `Param validation error: ${error}`;
+    if (hasKey(values, paramName) && hasKey(param, "validation")) {
+      const error = param.validation(values[param]);
+      if (error !== null && error !== undefined) {
+        return `Param validation error returned: ${error}`;
       }
     }
   }
 
   // Check that there are not other values besides parameters and default params
-  for (let param in values) {
-    if (!hasKey(params, param) && !hasKey(otherParams, param)) {
-      return `Unexpected ${param} value not defined on template.`;
+  for (let paramName in values) {
+    if (!hasKey(params, paramName) && !hasKey(otherParams, paramName)) {
+      return `Unexpected ${paramName} value not defined on template.`;
     }
   }